requirements.txt
fastapi==0.112.0
uvicorn[standard]==0.30.6
mangum==0.17.0
psycopg2-binary==2.9.9
openai==1.40.0
boto3==1.34.162
python-dotenv==1.0.1